Have added the drop down navigation menu.  Currently implemented with multiple activities, but this is not the desired effect so I need to look at how I might be able to convert to a single activity with different fragments.  Probably will involve a lot of hair pulling.  (All of this for a navigation mode that is deprecated in 21, but still seems the most logical way to implement the desired effect because the drawer isn't going to cut it and I already have swiping tabs underneath. Sigh.)
Known crash associated with the shopping list functionality, where quantities of thread required for the pattern are not being stored properly and resulting in a null pointer exception.  Stash clearing functionality updated to take into account the additional lists and clear out that NPE.  I am probably trying to do too many things at once but at least there is some ability to test the code for crashes, even if things are not right.

Embellishments are not properly reflecting what is owned versus not owned, although threads are and displaying properly in master versus stash lists.  At least something works.