Major revamp of the thread fragment.  Added display of pattern list, ability to manually add skeins to the shopping list (with a display that hopefully makes it clear what comes from kitted patterns and what has been added manually).  Changed display lists to show the total for purchase, with breakdown on the individual tab.  Managed to get the empty view for the list figured out, so need to go back and add that to several of the other listviews.  Also fixed it so that the lists are sorted pretty much everywhere, including paging through, as would be expected.  Still need to add the save/load for the additional skeins for purchase.
I may want to think about adding a toggle to allow folks to treat the last skeins as whole instead of partial, so that if they want to buy a full skein for each kit as it is called for, that can be accommodated.  In the settings menu, I think, assuming I ever figure out how that works.

First up though is adding the same changes to the embellishment page as I just did for threads.